fmap: Fix pointer related casts
BUG=None
BRANCH=None
TEST=Compiles

Original-Change-Id: I3a747cb562e7390bb81eca874d6c5aaa54b81e6e
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/209337
Original-Tested-by: Furquan Shaikh <furquan@chromium.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Commit-Queue: Furquan Shaikh <furquan@chromium.org>
(cherry picked from commit 5d3aef321a9313719308909ec40fdad0ec631a9f)

Reviewed-on: http://review.coreboot.org/8222
Tested-by: build bot (Jenkins)

+++ src/vendorcode/google/chromeos/fmap.c
@@ -122,7 +122,7 @@ int find_fmap_entry(const char name[], void **pointer)
 #endif
 	}
 
-	*pointer = (void*) ((u32)base + area->offset);
+	*pointer = (void*) ((uintptr_t)base + area->offset);
 	printk(BIOS_DEBUG, "FMAP: %s at %p (offset %x)\n",
 	       name, *pointer, area->offset);
 	return area->size;
